GUI ready and works (for as far as I can see) there are windows created on is with an overview of all the trips. The other has 3 tabs: -ItineraryPanel: add and remove activities during the trip -TravellerPanel: add and remove travellers from the trip -DayOverviewPanel: gives an overview for a selected day of all the activities that will happen that day
I've made small changes to some of the classes mostly adding some methods so we can see more in the GUI.
When double clicking an activity/traveller or trip in the GUI you get more detailed information about what you've clicked.

Possible improvements:
-more catches => check if the trip isn't planned in the past for example
-Give a more personal look to the GUI because now it's just a simple basic look => not necessarily a requirement
-add extra features maybe??

Important: the way the date is handled is in the format YYYY-MM-DD, i've tried working with a different format but it didn't worked YYYY-MM-DD is java's default and an industry standard. Normally i've put this format everywhere when we ask the user to input a date somewhere.